Upload OriginalBopItIdea.py
After the motors were too weak for my original idea, I decided to rebuild the "Bop it" game, where each corner of the platform corresponded to a different actuator: Twist it, Flick it, Spin it, and Pull it. The platform would tilt the ball into each of the different corners, as determined by a random number generator. Additionally, for each of the numbers, the EV3 would pull a trivia fact about the number via an API.